refactor: unify ownership managers
